[Model] Complete Echo-WM generation and precision fixes
Complete native checkpoint discovery, strict weight loading, image/text/action
conditioning, audiovisual decoding, and the stepwise runner lifecycle.
Correct Ulysses cache/RoPE layouts and preserve the reference dtype, geometry,
attention-kernel, and tensor-layout conventions. Add regression coverage,
offline generation examples, deployment defaults, and documentation.

Validation: 468 CPU tests passed; both single-GPU demonstration MP4s match
their same-runtime upstream counterparts byte-for-byte. TP/SP remain
experimental because long-rollout numerical parity is not yet accepted.

+++ vllm_omni/diffusion/models/echo_wm/causal_cache.py
@@ -344,7 +344,9 @@ def build_audio_positions(
         mel = (mel + 1 - latent_downsample).clamp(min=0)
         return mel * hop_length / sample_rate
 
-    latent_index = torch.arange(num_frames, device=device)
+    # The reference starts with FP32 indices. Integer true-division uses
+    # different rounding from its floating-point timestamp arithmetic.
+    latent_index = torch.arange(num_frames, device=device, dtype=torch.float32)
     start = frame_time(latent_index)
     end = frame_time(latent_index + 1)
     coords = torch.stack((start, end), dim=-1).float()  # (T, 2)
